# swagger/docs/swagger.yaml
basePath: /
definitions:
  main.HTTPError:
    description: HTTPError information
    properties:
      businessErrorCode:
        type: integer
      description:
        type: string
    type: object
  main.Pet:
    description: Pet information
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
host: localhost:1323
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Echo Swagger Example API
  version: "1.0"
paths:
  /pets:
    get:
      consumes:
      - application/json
      description: list pets
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main.Pet'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/main.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.HTTPError'
      summary: List of all pet
      tags:
      - pet
    post:
      consumes:
      - application/json
      description: Create a new pet with the given information
      parameters:
      - description: Pet object that needs to be added to the store
        in: body
        name: pet
        required: true
        schema:
          $ref: '#/definitions/main.Pet'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/main.Pet'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.HTTPError'
      summary: Create a new pet
      tags:
      - pet
schemes:
- http
swagger: "2.0"